+++ client/ui/modules/Library/LibraryMainFrameModule.java
@@ -1,4 +1,4 @@
-package client.ui.modules;
+package client.ui.modules.Library;
 
 import client.controller.LibraryController;
 import common.vo.UserVO;
